Added AttributedParser, moved around resource directories.
Added AttributedParser, which calls the Parser to obtain tokens, but uses a `Theme` to return attributes that can be applied to an NSAttributedString. I've moved the other.py.patch into a resource directory for PygmentsKit, since this is the only target that embeds pygments. Furthermore I've also improved the FontStyle loading support, although I have not tested this yet. Yet to improve is support for more scope selectors as it is also possible to separate them with spaces.